Fix memory leak with long-running contiguous event loop.
In fetch_events, @conn.garbage_collect only occurred before the
loop began, but not at any point during the loop. This means that
in an application that consists of one complete contiguous event
loop that never stops looping until the end of the program,
memory will accumulate in the underlying C library withou
ever being garbage collected. This change causes garbage
collection in the C library to occur before each event is fetched.