Add example programs for demonstrating memory use over time.
These examples are good demonstrations of whether the
exhibited use patterns are free of memory leaks. If so,
the two programs can be run together in perpetuity
and memory use will not grow. If memory use does grow
in one or both of the programs, there is a leak that
can be pinpointed by further reducing the case studies
and/or using other tools like valgrind.