Compatibility with Bundler and Ruby 2.4.
The generated code (under ffi/gen) has to be included, the code is never generated if installed by bundler.

Also removed the FFI requirement in `ext/rabbitmq/Rakefile` as FFI might not be available yet when the Rakefile is executed (in a clean bundle scenario), reimplemeneted the only function used from it instead. 

Also made it 2.4 compatible, removing the warning for Fixnum, while still being backward compatible.